requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
endee
sentence-transformers==2.7.0
python-dotenv==1.0.1
google-generativeai==0.7.2
python-multipart==0.0.9
